practice_7/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6